# backend/JobLog/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Job Logging
  version: 1.0.0
  description: Job Logging End Points
servers:
  - url: http://localhost:8080

paths:
  /user/{userName}:
    get:
      summary: Fetch user details
      description: This api fetches the user details for the provided username(email)
      operationId: getUser
      tags:
        - user
      parameters:
        - name: userName
          in: path
          required: true
          description: The username used to fetch user details
          schema:
            type: string
      responses:
        200:
          description: User details fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailsDTO'
        404:
          description: User not found
  /user:
    post:
      summary: Signup
      description: This API is used to create user
      operationId: createUser
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDetailsDTO'
      responses:
        200:
          description: User details created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: name of the user
        400:
          description: user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionHandlerDTO'

components:
  schemas:
    UserDetailsDTO:
      type: object
      properties:
        name:
          type: string
        email:
          description: Email of the user
          type: string
        phoneNumber:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        dateOfBirth:
          type: string
          format: date
    LoginDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    ExceptionHandlerDTO:
      type: object
      properties:
        status:
          type: int
        message:
          type: string
        timestamp:
          type: long